# openapi.yaml
openapi: "3.0.3"
info:
  title: Bowling Group Score Calculator API
  version: "1.0.0"
  description: >
    REST API for tracking bowling sessions, frame-by-frame scoring, and group leaderboards.
    All endpoints except /api/auth/login and /api/auth/register require a Bearer JWT token
    obtained from POST /api/auth/login.
servers:
  - url: http://localhost:8080
    description: Local development server
security:
  - bearerAuth: []

# ── Reusable components ───────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
          example: admin
          description: Registered username. Must not be blank.
        password:
          type: string
          minLength: 1
          example: changeme
          description: Account password. Must not be blank.
    RegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          example: alice
          description: Unique login name (3–50 characters).
        email:
          type: string
          format: email
          example: alice@example.com
          description: Valid email address. Must not be blank.
        password:
          type: string
          minLength: 6
          example: secret123
          description: Account password. Minimum 6 characters.
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          minLength: 1
          description: The user's current password. Used to verify identity before the change.
        newPassword:
          type: string
          minLength: 8
          description: Replacement password. Minimum 8 characters.
    CreateBowlerRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: Alice
          description: Bowler's display name (1–100 characters). Must not be blank.
    UpdateBowlerRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: Alice B.
          description: Bowler's new display name (1–100 characters). Must not be blank.
...
    TokenResponse:
      type: object
      properties:
        token:
          type: string
          description: 'JWT access token. Include in subsequent requests as ''Authorization: Bearer <token>''.'
        type:
          type: string
          example: Bearer
        username:
          type: string
          example: alice
        roles:
          type: array
          items:
            type: string
          example: [ROLE_USER]
    Bowler:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Alice
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "Session not found: 42"
    ValidationErrorResponse:
      type: object
      description: Returned when @Valid constraints are violated. Lists every failing field.
      properties:
        message:
          type: string
          example: Validation failed
        errors:
          type: object
          additionalProperties:
            type: string
          example:
            name: "Name must not be blank"
            sessionDate: "Session date is required"

# ── Paths ─────────────────────────────────────────────────────────────────────
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Obtain a JWT token
      description: No prior authentication required.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          description: JWT issued successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        "400":
          description: Blank username or password (field validation failure).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        "401":
          description: Credentials do not match any account.
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user account
      description: No prior authentication required. New accounts receive ROLE_USER.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        "200":
          description: Account created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User registered successfully
        "400":
          description: Validation error, or username/email already in use.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ValidationErrorResponse'
                  - $ref: '#/components/schemas/ErrorResponse'
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Issue a new JWT with reset expiry
      description: >
        Supply the current (still-valid) Bearer token. Returns a fresh token without
        requiring the user to re-enter their credentials.
      responses:
        "200":
          description: New token issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        "401":
          description: Missing or expired token.
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Get the authenticated user's profile
      responses:
        "200":
          description: Username and granted roles.
          content:
            application/json:
              schema:
                type: object
                properties:
                  username:
                    type: string
                    example: alice
                  roles:
                    type: array
                    items:
                      type: string
                    example: [ROLE_USER]
        "401":
          description: Missing or invalid JWT.
  /api/auth/change-password:
    post:
      tags: [Auth]
      summary: Change the authenticated user's password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        "200":
          description: Password updated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Password changed successfully
        "400":
          description: Validation failure or incorrect current password.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ValidationErrorResponse'
                  - $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Missing or invalid JWT.

  # ── Bowlers ───────────────────────────────────────────────────────────────
  /api/bowlers:
    get:
      tags: [Bowlers]
      summary: List all bowlers
      responses:
        "200":
          description: Full list of bowlers.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Bowler'
        "401":
          description: Missing or invalid JWT.

    post:
      tags: [Bowlers]
      summary: Create a bowler
      description: Requires ROLE_ADMIN. Name must be unique.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBowlerRequest'
      responses:
        "200":
          description: Bowler created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bowler'
        "400":
          description: Validation failure or duplicate name.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ValidationErrorResponse'
                  - $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Missing or invalid JWT.
        "403":
          description: Requires ROLE_ADMIN.
  /api/bowlers/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
        description: Bowler ID.

    get:
      tags: [Bowlers]
      summary: Get a bowler by ID
      responses:
        "200":
          description: Bowler found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bowler'
        "400":
          description: Bowler not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Missing or invalid JWT.

    put:
      tags: [Bowlers]
      summary: Update a bowler's name
      description: Requires ROLE_ADMIN.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBowlerRequest'
      responses:
        "200":
          description: Bowler updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bowler'
        "400":
          description: Validation failure or bowler not found.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ValidationErrorResponse'
                  - $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Missing or invalid JWT.
        "403":
          description: Requires ROLE_ADMIN.

    delete:
      tags: [Bowlers]
      summary: Delete a bowler
      description: >
        Requires ROLE_ADMIN. Rejected with 400 if the bowler has any existing game
        history — delete the session(s) first or transfer data before removing the bowler.
      responses:
        "204":
          description: Bowler deleted.
        "400":
          description: Bowler not found, or bowler has existing game history.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Missing or invalid JWT.
        "403":
          description: Requires ROLE_ADMIN.

  # ── Sessions ──────────────────────────────────────────────────────────────
...